requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
biopython=1.74=py37h516909a_0
bzip2=1.0.8=h516909a_1
ca-certificates=2019.9.11=hecc5488_0
certifi=2019.9.11=py37_0
curl=7.65.3=hf8cf82a_0
htslib=1.9=h4da6232_3
krb5=1.16.3=h05b26f9_1001
lastz=1.0.4=h14c3975_3
libblas=3.8.0=14_openblas
libcblas=3.8.0=14_openblas
libcurl=7.65.3=hda55be3_0
libdeflate=1.3=h516909a_0
libedit=3.1.20170329=hf8c457e_1001
libffi=3.2.1=he1b5a44_1006
libgcc-ng=9.1.0=hdf63c60_0
libgfortran-ng=7.3.0=hdf63c60_2
liblapack=3.8.0=14_openblas
libopenblas=0.3.7=h6e990d7_2
libssh2=1.8.2=h22169c7_2
libstdcxx-ng=9.1.0=hdf63c60_0
ncurses=6.1=hf484d3e_1002
numpy=1.17.2=py37h95a1406_0
openssl=1.1.1c=h516909a_0
pip=19.3.1=py37_0
python=3.7.3=h33d41f4_1
readline=8.0=hf8c457e_0
samtools=1.9=h10a08f8_12
setuptools=41.4.0=py37_0
sqlite=3.30.1=hcee41ef_0
tk=8.6.9=hed695b0_1003
wheel=0.33.6=py37_0
xz=5.2.4=h14c3975_1001
zlib=1.2.11=h516909a_1006
